+++ src/main/kotlin/com/gmaur/investment/r4automator/app/R4FundsParser.kt
@@ -1,12 +1,9 @@
 package com.gmaur.investment.r4automator.app
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.databind.SerializationFeature
-import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import com.gmaur.investment.r4automator.infrastructure.files.FilePickerProvider
 import com.gmaur.investment.r4automator.infrastructure.files.FileUtils
+import com.gmaur.investment.r4automator.infrastructure.funds.FilePortfolioRepository
 import com.gmaur.investment.r4automator.infrastructure.funds.ParseFunds
-import com.gmaur.investment.r4automator.infrastructure.portfolio.Portfolio
 import java.io.File
 import java.nio.file.Paths
 
@@ -25,17 +22,4 @@ class R4FundsParser(private val portfolioRepo: FilePortfolioRepository) {
 }
 fun main(args: Array<String>) {
     R4FundsParser(FilePortfolioRepository()).run(args)
-}
-
-
-class FilePortfolioRepository {
-
-    private var mapper: ObjectMapper = ObjectMapper().registerKotlinModule()
-
-
-    fun save(portfolio: Portfolio, file: File) {
-        mapper.enable(SerializationFeature.INDENT_OUTPUT)
-        mapper.writeValue(file, portfolio)
-    }
-
 }
